Fixed a small bug when adding the same product to the basket was merging a quantity. Also, changed parseInt to parseFloat on basket summary, so that a correct price is displayed

+++ src/components/Basket/Basket.js
@@ -20,7 +20,7 @@ const Basket = () => {
                         </div>
                         <div className="basket__summary">
                             <h2>Order Summary</h2>
-                            <p>total: £{parseInt(basket.total).toFixed(2)}</p>
+                            <p>total: £{parseFloat(basket.total).toFixed(2)}</p>
                             <button>Checkout</button>
                         </div>
                     </>
